Move common_transcript from mng_claude_mind plugin into mng claude plugin
The common_transcript functionality converts raw Claude transcript events
into an agent-agnostic format. This was previously a Python watcher in the
mng_claude_mind plugin, but since all Claude agents should be able to emit
common transcripts, it belongs in the base claude plugin.

Changes:
- Add emit_common_transcript config option to ClaudeAgentConfig (default: false)
- Create common_transcript.sh bash script that watches stream_transcript
  output and converts to common format
- Launch common_transcript.sh from claude_background_tasks.sh when
  MNG_EMIT_COMMON_TRANSCRIPT=1
- Write events to events/claude/common_transcript/events.jsonl
- ClaudeMindConfig defaults emit_common_transcript to true
- Delete transcript_watcher.py, transcript_watcher.sh, and associated
  CLI command/tmux window from mng_claude_mind plugin
