+++ libs/mng_claude_mind/imbue/mng_claude_mind/plugin.py
@@ -48,9 +48,6 @@ CONV_WATCHER_COMMAND: Final[str] = "mng mindconversations"
 EVENT_WATCHER_WINDOW_NAME: Final[str] = "events"
 EVENT_WATCHER_COMMAND: Final[str] = "mng mindevents"
 
-TRANSCRIPT_WATCHER_WINDOW_NAME: Final[str] = "transcript"
-TRANSCRIPT_WATCHER_COMMAND: Final[str] = "mng mindtranscript"
-
 # Web server: serves the main web interface with conversation selector
 # and agent list page.
 WEB_SERVER_WINDOW_NAME: Final[str] = "web_server"
@@ -69,6 +66,11 @@ class ClaudeMindConfig(ClaudeAgentConfig):
         default=True,
         description="Automatically trust the agent's working directory in ~/.claude.json. "
         "Enabled by default for minds since they run in-place in their own repo.",
+    )
+    emit_common_transcript: bool = Field(
+        default=True,
+        description="Emit a common, agent-agnostic transcript alongside the raw Claude transcript. "
+        "Enabled by default for minds so that other watchers can consume structured events.",
     )
     install_llm: bool = Field(
         default=True,
@@ -97,7 +99,6 @@ class ClaudeMindAgent(ClaudeAgent):
     - Conversation watcher (syncs llm DB to events/messages/events.jsonl)
     - Event watcher (sends new events to primary role agent via mng message)
     - Web server (main web interface with conversation selector and agent list)
-    - Transcript watcher (converts claude_transcript to common_transcript)
     """
 
     enter_submission_timeout_seconds: float = Field(
@@ -288,7 +289,9 @@ def inject_supporting_services(params: dict[str, Any]) -> None:
     - Conversation watcher (syncs llm DB to JSONL files)
     - Event watcher (sends new events to primary role agent via mng message)
     - Web server (main web interface with conversation selector and agent list)
-    - Transcript watcher (converts claude_transcript to common_transcript)
+
+    Note: transcript conversion (common_transcript) is now handled by the
+    claude plugin's background tasks script, controlled by emit_common_transcript.
 
     Agent and chat terminal access is provided by the consolidated ttyd
     server (from mng_ttyd) via URL-arg dispatch to commands/ttyd/*.sh scripts.
@@ -299,7 +302,6 @@ def inject_supporting_services(params: dict[str, Any]) -> None:
         f'{CONV_WATCHER_WINDOW_NAME}="{CONV_WATCHER_COMMAND}"',
         f'{EVENT_WATCHER_WINDOW_NAME}="{EVENT_WATCHER_COMMAND}"',
         f'{WEB_SERVER_WINDOW_NAME}="{WEB_SERVER_COMMAND}"',
-        f'{TRANSCRIPT_WATCHER_WINDOW_NAME}="{TRANSCRIPT_WATCHER_COMMAND}"',
     )
 
 
@@ -340,8 +342,8 @@ def override_command_options(
 ) -> None:
     """Add mind supporting service windows when creating claude-mind role agents (or subtypes).
 
-    Injects: conversation watcher, event watcher, web server,
-    and transcript watcher as supporting services.
+    Injects: conversation watcher, event watcher, and web server
+    as supporting services.
 
     Matches any agent type whose registered class is ClaudeMindAgent or
     a subclass of it (e.g. elena-code, custom mind types).
