+++ libs/mng_claude_mind/imbue/mng_claude_mind/provisioning_test.py
@@ -1,12 +1,7 @@
 """Unit tests for the mng_claude_mind provisioning module."""
 
 import importlib
-import json
 import os
-import subprocess
-import sys
-import tempfile
-import types
 from pathlib import Path
 from typing import Any
 from typing import cast
@@ -17,7 +12,6 @@ from imbue.mng_claude_mind.conftest import StubCommandResult
 from imbue.mng_claude_mind.conftest import StubHost
 from imbue.mng_claude_mind.conftest import create_mind_conversations_table_in_test_db
 from imbue.mng_claude_mind.conftest import write_conversation_to_db
-from imbue.mng_claude_mind.data_types import CommonToolResultEvent
 from imbue.mng_claude_mind.data_types import ProvisioningSettings
 from imbue.mng_claude_mind.provisioning import MIND_CONVERSATIONS_TABLE_SQL
 from imbue.mng_claude_mind.provisioning import TalkingRoleConstraintError
@@ -52,298 +46,6 @@ def test_load_mind_resource_loads_resource() -> None:
     """Check that the load_mind_resource works at all"""
     content = load_mind_resource("chat.sh")
     assert "#!/bin/bash" in content
-
-
-# -- Transcript watcher conversion logic tests --
-
-
-# FIXME: transcript_watcher.sh is entirely dead code and should be entirely removed... what are these tests doing?
-def _extract_convert_script() -> str:
-    """Extract the inline Python CONVERT_SCRIPT from transcript_watcher.sh."""
-    content = load_mind_resource("transcript_watcher.sh")
-    start_marker = "python3 << 'CONVERT_SCRIPT'"
-    start_idx = content.index(start_marker)
-    start_of_python = content.index("\n", start_idx) + 1
-    remaining = content[start_of_python:]
-    python_lines = []
-    for line in remaining.split("\n"):
-        if line.strip() == "CONVERT_SCRIPT":
-            break
-        python_lines.append(line)
-    return "\n".join(python_lines)
-
-
-def _run_conversion(
-    input_lines: list[str],
-    existing_output_lines: list[str] | None = None,
-    tmp_path: Path = Path("/tmp"),
-) -> list[dict[str, Any]]:
-    """Run the conversion logic on the given input and return output events."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        input_file = Path(tmpdir) / "input.jsonl"
-        output_file = Path(tmpdir) / "output.jsonl"
-
-        input_file.write_text("\n".join(input_lines) + "\n" if input_lines else "")
-
-        if existing_output_lines:
-            output_file.write_text("\n".join(existing_output_lines) + "\n")
-
-        script = _extract_convert_script()
-        env = {
-            **os.environ,
-            "_INPUT_FILE": str(input_file),
-            "_OUTPUT_FILE": str(output_file),
-        }
-
-        result = subprocess.run(
-            ["python3", "-c", script],
-            capture_output=True,
-            text=True,
-            env=env,
-            timeout=10,
-        )
-        if result.returncode != 0:
-            raise RuntimeError(f"Conversion failed: {result.stderr}")
-
-        if not output_file.exists():
-            return []
-
-        output_text = output_file.read_text()
-        events = []
-        for line in output_text.strip().split("\n"):
-            if line.strip():
-                events.append(json.loads(line))
-        return events
-
-
-def test_conversion_handles_user_text_message() -> None:
-    raw = json.dumps(
-        {
-            "type": "user",
-            "uuid": "user-uuid-1",
-            "timestamp": "2026-01-01T00:00:00Z",
-            "message": {"role": "user", "content": "Hello world"},
-        }
-    )
-    events = _run_conversion([raw])
-    assert len(events) == 1
-    assert events[0]["type"] == "user_message"
-    assert events[0]["content"] == "Hello world"
-    assert events[0]["source"] == "common_transcript"
-    assert events[0]["event_id"] == "user-uuid-1-user"
-
-
-def test_conversion_handles_assistant_message_with_text() -> None:
-    raw = json.dumps(
-        {
-            "type": "assistant",
-            "uuid": "asst-uuid-1",
-            "timestamp": "2026-01-01T00:00:01Z",
-            "message": {
-                "role": "assistant",
-                "model": "claude-opus-4.6",
-                "content": [{"type": "text", "text": "Hello back!"}],
-                "stop_reason": "end_turn",
-                "usage": {"input_tokens": 100, "output_tokens": 50},
-            },
-        }
-    )
-    events = _run_conversion([raw])
-    assert len(events) == 1
-    assert events[0]["type"] == "assistant_message"
-    assert events[0]["model"] == "claude-opus-4.6"
-    assert events[0]["text"] == "Hello back!"
-
-
-def test_conversion_handles_tool_results() -> None:
-    assistant = json.dumps(
-        {
-            "type": "assistant",
-            "uuid": "asst-uuid-3",
-            "timestamp": "2026-01-01T00:00:03Z",
-            "message": {
-                "role": "assistant",
-                "model": "claude-opus-4.6",
-                "content": [
-                    {"type": "tool_use", "id": "toolu_456", "name": "Read", "input": {"file_path": "/tmp/test.txt"}},
-                ],
-                "stop_reason": "tool_use",
-                "usage": {"input_tokens": 50, "output_tokens": 30},
-            },
-        }
-    )
-    user_result = json.dumps(
-        {
-            "type": "user",
-            "uuid": "user-uuid-2",
-            "timestamp": "2026-01-01T00:00:04Z",
-            "message": {
-                "role": "user",
-                "content": [
-                    {
-                        "type": "tool_result",
-                        "tool_use_id": "toolu_456",
-                        "content": "file contents here",
-                        "is_error": False,
-                    },
-                ],
-            },
-        }
-    )
-    events = _run_conversion([assistant, user_result])
-    tool_results = [e for e in events if e["type"] == "tool_result"]
-    assert len(tool_results) == 1
-    assert tool_results[0]["tool_call_id"] == "toolu_456"
-    assert tool_results[0]["tool_name"] == "Read"
-
-
-def test_conversion_skips_progress_events() -> None:
-    raw = json.dumps(
-        {
-            "type": "progress",
-            "uuid": "prog-uuid-1",
-            "timestamp": "2026-01-01T00:00:05Z",
-            "data": {"type": "bash_progress"},
-        }
-    )
-    events = _run_conversion([raw])
-    assert len(events) == 0
-
-
-def test_conversion_deduplicates_by_event_id() -> None:
-    raw = json.dumps(
-        {
-            "type": "user",
-            "uuid": "user-uuid-3",
-            "timestamp": "2026-01-01T00:00:07Z",
-            "message": {"role": "user", "content": "dedup test"},
-        }
-    )
-    existing = json.dumps(
-        {
-            "timestamp": "2026-01-01T00:00:07Z",
-            "type": "user_message",
-            "event_id": "user-uuid-3-user",
-            "source": "common_transcript",
-            "role": "user",
-            "content": "dedup test",
-        }
-    )
-    events = _run_conversion([raw], existing_output_lines=[existing])
-    assert len(events) == 1  # only the pre-existing one
-
-
-def test_conversion_handles_malformed_lines_gracefully() -> None:
-    valid = json.dumps(
-        {
-            "type": "user",
-            "uuid": "user-uuid-4",
-            "timestamp": "2026-01-01T00:00:08Z",
-            "message": {"role": "user", "content": "valid"},
-        }
-    )
-    events = _run_conversion(["not json at all", valid, '{"incomplete": true'])
-    assert len(events) == 1
-    assert events[0]["content"] == "valid"
-
-
-def test_conversion_user_message_validates_against_pydantic_schema() -> None:
-    pass
-
-
-# -- Transcript watcher conversion logic tests --
-
-
-def _strip_below_watchdog_marker(source: str) -> str:
-    """Strip everything at and below the WATCHDOG-DEPENDENT marker line.
-
-    Searches for the marker as a standalone comment line (not inside a string
-    literal like a docstring). Returns the source text up to (but not including)
-    the marker line.
-    """
-    marker_prefix = "# --- WATCHDOG-DEPENDENT CODE BELOW"
-    lines = source.split("\n")
-    # Walk backwards to find the last occurrence (the real one, not a docstring mention)
-    for i in range(len(lines) - 1, -1, -1):
-        if lines[i].startswith(marker_prefix):
-            return "\n".join(lines[:i])
-    raise ValueError(f"Marker {marker_prefix!r} not found in source")
-
-
-def test_conversion_assistant_message_validates_against_pydantic_schema() -> None:
-    pass
-
-
-def _load_transcript_watcher_module() -> dict[str, Any]:
-    """Load the stdlib-only portion of transcript_watcher.py (above the watchdog marker).
-
-    Returns a namespace dict containing the conversion functions.
-    """
-    content = load_mind_resource("transcript_watcher.py")
-    # Also load watcher_common.py (stripped above its watchdog marker) so
-    # transcript_watcher.py can import Logger from it.
-    watcher_common_content = load_mind_resource("watcher_common.py")
-    watcher_common_stripped = _strip_below_watchdog_marker(watcher_common_content)
-
-    stripped = _strip_below_watchdog_marker(content)
-
-    # Exec watcher_common first so transcript_watcher can import from it
-    watcher_common_ns: dict[str, Any] = {}
-    exec(compile(watcher_common_stripped, "watcher_common.py", "exec"), watcher_common_ns)
-
-    # Patch sys.modules so `from watcher_common import ...` works during exec
-    watcher_common_mod = types.ModuleType("watcher_common")
-    for key, value in watcher_common_ns.items():
-        if not key.startswith("__"):
-            setattr(watcher_common_mod, key, value)
-
-    old_mod = sys.modules.get("watcher_common")
-    sys.modules["watcher_common"] = watcher_common_mod
-    try:
-        ns: dict[str, Any] = {"__file__": "/tmp/transcript_watcher.py"}
-        exec(compile(stripped, "transcript_watcher.py", "exec"), ns)
-    finally:
-        if old_mod is not None:
-            sys.modules["watcher_common"] = old_mod
-        else:
-            sys.modules.pop("watcher_common", None)
-    return ns
-
-
-def test_conversion_tool_result_validates_against_pydantic_schema() -> None:
-    assistant = json.dumps(
-        {
-            "type": "assistant",
-            "uuid": "contract-asst-2",
-            "timestamp": "2026-01-01T00:00:02.000000000Z",
-            "message": {
-                "role": "assistant",
-                "model": "claude-opus-4.6",
-                "content": [{"type": "tool_use", "id": "toolu_c2", "name": "Read", "input": {"file": "test.txt"}}],
-                "stop_reason": "tool_use",
-                "usage": {"input_tokens": 50, "output_tokens": 30},
-            },
-        }
-    )
-    user_result = json.dumps(
-        {
-            "type": "user",
-            "uuid": "contract-user-2",
-            "timestamp": "2026-01-01T00:00:03.000000000Z",
-            "message": {
-                "role": "user",
-                "content": [
-                    {"type": "tool_result", "tool_use_id": "toolu_c2", "content": "file contents", "is_error": False}
-                ],
-            },
-        }
-    )
-    events = _run_conversion([assistant, user_result])
-    tool_results = [e for e in events if e["type"] == "tool_result"]
-    assert len(tool_results) == 1
-    validated = CommonToolResultEvent.model_validate(tool_results[0])
-    assert validated.tool_call_id == "toolu_c2"
-    assert validated.tool_name == "Read"
 
 
 # -- Memory linker content tests --
@@ -615,7 +317,6 @@ def test_create_event_log_directories_creates_all_source_dirs() -> None:
         "stop",
         "monitor",
         "delivery_failures",
-        "common_transcript",
         "servers",
     ):
         assert any(source in c and "mkdir" in c for c in host.executed_commands), f"Missing mkdir for {source}"
