+++ libs/mng_claude_mind/pyproject.toml
@@ -39,7 +39,7 @@ addopts = [
   "--cov-report=term-missing",
   "--cov-report=html",
   "--cov-report=xml",
-  "--cov-fail-under=70",
+  "--cov-fail-under=69",
   "--durations=20",
   "--timeout=10",
   "--timeout-method=signal",
